podlewajnik_backend/pyproject.toml
[tool.poetry]
name = "podlewajnik"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "podlewajnik", from="src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
tortoise-orm = "^0.21.3"
aerich = "^0.7.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
pyright = "^1.1.368"
rich = "^13.7.1"

[tool.aerich]
tortoise_orm = "podlewajnik.database.config.TORTOISE_ORM"
location = "./src/podlewajnik/database/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
